App
Demonstrating the purposes and uses of encapsulation: a core practice of programming. All code should be made private if possible. The only things public should be methods that are called by the end user. E.g. getters and setters such as getEvilNumber() or setEvilNumber(). Practicing encapsulation ensures efficient flow and minimizes issues.